Update cost field's dataType
- Changed cost field type from Integer to Float in OpenAPI spec.
- Updated cost field to BigDecimal in DTOs and domain models.
- Added validation for minimum value and maximum 2 decimal places.
- Updated tests to reflect new constraints.

+++ src/main/kotlin/com/store/domain/Product.kt
@@ -2,13 +2,14 @@ package com.store.domain
 
 import com.store.dto.ProductRequest
 import com.store.enums.ProductType
+import java.math.BigDecimal
 
 data class Product(
     val id: Int,
     val name: String,
     val type: ProductType,
     val inventory: Int,
-    val cost: Int
+    val cost: BigDecimal
 ) {
     constructor(productId: ProductId, productRequest: ProductRequest) : this(
         id = productId.id,
